Make test processor work
